// src/app/core/api/gateway-types.ts
export interface components {
    schemas: {
        /** ChatCompletionRequest */
        ChatCompletionRequest: {
            /** Max Tokens */
            max_tokens?: number | null;
            /** Messages */
            messages: components["schemas"]["ChatMessage"][];
            /** Model */
            model: string;
            /**
             * Stream
             * @default false
             */
            stream: boolean;
            /** Temperature */
            temperature?: number | null;
        };
        /** ChatMessage */
        ChatMessage: {
            /** Content */
            content: string;
            /** Role */
            role: string;
        };
        /** EmbeddingData */
        EmbeddingData: {
            /** Embedding */
            embedding: number[];
            /** Index */
            index: number;
            /**
             * Object
             * @default embedding
             */
            object: string;
        };
        /** EmbeddingRequest */
        EmbeddingRequest: {
            /** Input */
            input: string | string[];
            /** Model */
            model: string;
        };
        /** EmbeddingResponse */
        EmbeddingResponse: {
            /** Data */
            data: components["schemas"]["EmbeddingData"][];
            /** Model */
            model: string;
            /**
             * Object
             * @default list
             */
            object: string;
            usage: components["schemas"]["EmbeddingUsage"];
        };
        /** EmbeddingUsage */
        EmbeddingUsage: {
            /** Prompt Tokens */
            prompt_tokens: number;
            /** Total Tokens */
            total_tokens: number;
        };
        /** HTTPValidationError */
        HTTPValidationError: {
            /** Detail */
            detail?: components["schemas"]["ValidationError"][];
        };
        /** Model */
        Model: {
            /** Id */
            id: string;
            /**
             * Object
             * @default model
             */
            object: string;
            /** Owned By */
            owned_by: string;
        };
        /** ModelList */
        ModelList: {
            /** Data */
            data: components["schemas"]["Model"][];
            /**
             * Object
             * @default list
             */
            object: string;
        };
        /** ValidationError */
        ValidationError: {
            /** Location */
            loc: (string | number)[];
            /** Message */
            msg: string;
            /** Error Type */
            type: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}